+++ CHANGELOG.md
@@ -64,6 +64,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Camera option to set the clip start (#484)
   - Traceback dump when Blender output contains it
   - Subdirectory for each output generated (#541)
+  - Option to disable the denoiser (#539)
 - KiKit
   - Expand text variables and KiBot %X markers in text objects (see #497)
 - PCB Print: